api reorganization
Closes #2434.
