+++ mesa/visualization/__init__.py
@@ -1,7 +1,9 @@
 """Solara based visualization for Mesa models."""
 
-from .components.altair import make_space_altair
-from .components.matplotlib import make_plot_component, make_space_component
+from .components.altair_components import make_space_altair
+from .components.matplotlib_components import make_plot_component, make_space_component, make_plot_measure, make_space_matplotlib
+from .components.mpl_space_drawing import (draw_space, draw_continuous_space, draw_network,
+                                           draw_orthogonal_grid, draw_voroinoi_grid, draw_hex_grid, draw_property_layers)
 from .solara_viz import JupyterViz, SolaraViz
 from .UserParam import Slider
